refactor(hooks): extract modal handlers from MainApp into useModalHandlers
